requirements.txt
numpy
pandas
requests
scikit-learn
nltk
flask
authlib
python-dotenv
gunicorn
firebase-admin
